Completed first MERN app. Pending: Hosting on Vercel.

+++ backend/src/controllers/notes.ts
@@ -1,11 +1,16 @@
 import { RequestHandler } from "express";
-import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import NoteModel from "../models/note";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+
+        const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -14,8 +19,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
 
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Note ID is invalid!");
@@ -27,7 +34,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
             throw createHttpError(404, "Note ID not found!");
         }
 
-        res.status(200).json(note);        
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "Sorry, you are restricted from accessing this note!");
+        }
+
+        res.status(200).json(note);
     }
     catch (error) {
         next(error);
@@ -42,14 +53,17 @@ interface PostNoteBody {
 export const postNote: RequestHandler<unknown, unknown, PostNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
 
         if (!title) {
             throw createHttpError(400, "Title is required for a Note!");
         }
 
-        const createdNote = await NoteModel.create({ 
+        const createdNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             content: content,
         });
@@ -73,7 +87,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const updatedTitle = req.body.title;
     const updatedContent = req.body.content;
 
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Note ID is invalid!");
         }
@@ -86,6 +104,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note) {
             throw createHttpError(404, "Note ID not found!");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "Sorry, you are restricted from accessing this note!");
         }
 
         note.title = updatedTitle;
@@ -102,14 +124,23 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
 
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Note ID is invalid!");
         }
 
         const note = await NoteModel.findById(noteId).exec();
+
         if (!note) {
             throw createHttpError(404, "Note ID not found!");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "Sorry, you are restricted from accessing this note!");
         }
 
         await note.deleteOne();
